+++ packages/css-color-parser/src/functions/color-keyword.ts
@@ -1,9 +1,10 @@
 import type { ColorData } from '../color-data';
 import { ColorSpace } from '../color-space';
 import { SyntaxFlag } from '../color-data';
+import { toLowerCaseAZ } from '../util/to-lower-case-a-z';
 
 export function colorKeyword(name: string): ColorData | -1 {
-	if (name === 'transparent') {
+	if (toLowerCaseAZ(name) === 'transparent') {
 		return {
 			colorSpace: ColorSpace.XYZ_D50,
 			channels: [0, 0, 0],