+++ packages/media-query-list-parser/src/nodes/media-type.ts
@@ -1,4 +1,5 @@
 import { TokenIdent, TokenType } from '@csstools/css-tokenizer';
+import { toLowerCaseAZ } from '../util/to-lower-case-a-z';
 
 export enum MediaType {
 	/** Always matches */
@@ -28,7 +29,7 @@ export function typeFromToken(token: TokenIdent): MediaType | false {
 		return false;
 	}
 
-	const matchingValue = token[4].value.toLowerCase();
+	const matchingValue = toLowerCaseAZ(token[4].value);
 	switch (matchingValue) {
 		case MediaType.All:
 			return MediaType.All;